added some JavaDoc, refactored code, removed not needed exec variables
The id of the created QuestionnaireResponse is no longer set as a
'questionnaireResponseId' variable, since it is not needed for default
use cases. Process plugin creators can still register a variable with
the value by extending the DefaultUserTaskListener and overriding the
afterQuestionnaireResponseCreate(...) method.

The variable for the completed QuestionnaireResponse is now hidden and a
new method getLatestReceivedQuestionnaireResponse() has been added to
the Variables interface.

Removed badly named Target factory methods from the Variables interface
and added two new createTarget methods.

Added a new ListenerVariables interface to access the VariablesImpl
class via the Start-, Continue- and EndListener classes without
depending on the implementation. Creation of a VariablesImpl object now
via injected factory function.

Removed the TaskListener type parameter from the ProcessPlugin interface
and reworked the plugin spring context initialization. The
DefaultUserTaskListener is now available as a prototype bean.

JavaDoc for API service interfaces TBD.