new environment variable SERVER_CONTEXT_PATH default /fhir
The new environment variable SERVER_CONTEXT_PATH (default value "/fhir")
can be used to configure the reverse-proxy path that gets delegated to
the DSF FHIR app server. The context path of the DSF FHIR app server is
not modified by this variable.

Using `/fhir` as the target path works as long as the DSF FHIR app
server's config parameter `dev.dsf.server.context.path` is not modified
as well.