Fix paging logic in search result handling of resource provider
The results of a FHIR search query using search entry mode 'included'
could lead to infinite number of requests of incrementing result page
number if the number of resources of the join (with search mode
'included') in the results is less than the number of the primary
resources (with search mode 'match').
This fix only requests the next result page if the previous result
bundle contains a 'next' link.

Add test cases for OrganizationProviderImpl.