new dsf-maven plugin with multiple goals including help
* Goals are executed on the cli by using `dsf:` as a prefix, e.g `mvn
dsf:help`. All cli configurable parameters use a prefix `dsf.` e.g. `mvn
dev.dsf:dsf-maven-plugin:generate-config-doc
-Ddsf.configDocPackages=some.java.package`

* The `generate` goal from the old `dsf-tools-documentation-generator`
maven plugin is now part of this new maven plugin and renamed to
`generate-config-doc`. The old config parameter `workingPackages` was
renamed to `configDocPackages`. The goal can be used by v1 and v2
process plugins and is also configured for dsf-bpe-test-plugin-v1 and
dsf-bpe-test-plugin-v2 as well as dsf-bpe-server-jetty and
dsf-fhir-server-jetty. By default the goal executed in the
`process-classes` phase if an execution is configured in the pom.

* The goal `generate-dev-setup-cert-files` can be used to generate cert
files for a dev setup if configured in the pom and
`clean-dev-setup-cert-files` to remove the generated files. For the dsf
no execution is configured for both goals (circular dependency) and they
need to be executed from the cli using: `mvn
dsf:generate-dev-setup-files` and `mvn dsf:clean-dev-setup-files`. The
generate... goal stores certificates and private-keys in a `certDir`
folder and then copies files to configured targets. The "original" files
in the `certDir` can also be deleted by executing `mvn
dsf:clean-dev-setup-files -Ddsf.includeCertDir=true`

* Additional DSF specific goals include `generate-default-ca-files` to
generate the default Root CAs, Issuing CAs and CA chain files for the
default trusted CAs now stored in src/main/resource/cert, as well as
`generate-fhir-bundle` to generate the internal fhir bundle for the
dsf-fhir-validation module.

* A list of supported goals can be shown by running `mvn
dev.dsf:dsf-maven-plugin:help` for the latest version in the local .m2
repo or `mvn dev.dsf:dsf-maven-plugin:2.0.0:help` for a specific version
like 2.0.0 (not released yet).