modifications to support optional thumbprints on endpoint resource
- Identity classes now include a field for the associated Endpoint
resource. For a remote user authenticated by client certificate the code
tries to find the matching Endpoint resource by looking for an Endpoint
in the Database with a matching thumbprint, if no resource is found and
only one Endpoint exists for the matched organization, the matching
Endpoint resource is used. If there are multiple Endpoint resources for
the matched organization, no Endpoint is returned and the Task
authorization code falls back to the old behavior of using all
ActivityDefinition resources of the organization, to allow Task creates.
For the local client certificate or "practitioner" users the local
Endpoint based on the baseUrl address is used.
- When retrieving OrganizationAffiliation resources in the
TaskAuthorizationRule for the requester and recipient, the associated
Endpoint (if known) is used to query the DB more precisely.
- Plugin loading code modified to change the baseDefinition property of
StructureDefinition resource for Task profiles. The baseDefinition may
not include a version in order for the Base-Task profile to be updated
to version 2.0.0. Process plugins usually do not specify the Base-Task
version specific.
- OrganizationAffiliationHistoryIdentityFilter inheritance fix with no
change to the behavior.
- Some additional tests, code cleanup, incl. class to record
conversions.