compression service to use with InputStreams
Adds compression service with methods for gzip, bzip2 and lzma2. The
service is designed to be used with InputStreams. For the compression
methods this means that the incoming InputStream needs to be read,
converted to an OuputStream, used in the compression implementations and
than converted back to an InputStream.
Adds v2 plugin integration tests and a direct service test.