new service to modify start task output components
- Access to StartTaskUpdater service via Variables interface. The
implementation needs access to the start task process variable and thus
adding a getStartTaskUpdater() method to the Variables interface is the
solution with the smallest impact on existing interfaces. Alternatively
we could add the StartTaskUpdater as a third parameter to the
ServiceTask#execute(ProcessPluginApi, Variables) method or have the user
supply the start task manually to the StartTaskUpdater methods.