new oidc options to specify required aud and enable azp verification
- New required-audience yaml property can be used to specify a single
value or list of required audience values in the aud claim of the oidc
token.
- New verify-authorized-party (default true) can be used to disable the
verification of the authorized party value in the azp claim of the oidc
token. When enabled the azp claim is expected to be equal to the
client-id value.
- New bpe config property
dev.dsf.bpe.fhir.client.connections.config.default.enable.debug.logging
can be used to disable azp checks in general for all fhir client
connections that use oidc authentication. Per connection yaml property
verify-authorized-party has higher priority.
- Adds demo required-audience check to dic1-bpe in 3dic/ttp test setup.