enforced default profiles, fixed tests, improved ConformanceService
- Default profiles are now enforce during create and update operations
(single resources and via transaction/batch bundle).
- DefaultProfileProvider is used to add a default DSF profile to a new
or to be updated resource if the resource does not already declare a
profile supported as on of the default profiles. Some exceptions: Since
we do not have a default (non abstract) profile for Task resources, no
profile is added. But Task resource will need to have a specialized
profile declared to make it past the authorization rule. For
Organization resources, the member organization profile is considered
the default, parent organizations are allowed if the
.../organization-parent profile is self declared. The parent profile is
considered a "secondary" default profile. Not added by default, but no
other will be enforced if self declared.
- Fixed integration tests, where non valid resources where used.
- Improved ConformanceService: The CapabilityStatement resource now
declared DSF profiles as the rest.resource.profile property the the
exception of Task, still declaring the FHIR base profile. The parent
organization profile is declared as a rest.resource.supportedProfile
same is true for Task profiles from process plugins.