fix for process definition selection in old installs without tenant-id
Process models in DSF 1.x where deployed without the process API version
as tenant-id. If an process is updated on DSF 2.x it is deployed with
the process API version as tenant-id. The "getProcessDefinition" method
select the process definition with the highest internal version but
needs to sort by tenant-id first to preferentially select processes with
tenant-id.